Turn page.html into a base template
Moved the HTML boilerplate for all pages into page.html and put the
{{ content }} of page.html into the body as-is so that page.html can be
used as a base template instead of always having to include header.html
and footer.html in all pages. Also added page.html as the layout for
the project.html laout; this is how to inherit from page.html.