Switch fonts to Work Sans/Montserrat
Stick with two fonts for now because it's easier to make it work well.
Montserrat looks great for headings, and Work Sans is quite legible as
body text (since I'm not using italics anywhere). Also increased the
line height for Work Sans and set a max width on paragraphs for more
comfortable reading.